// src/dotnet/Blazor3D/Blazor3D.TypeScript/src/Generated/buffer-geometry.ts
/**
 * This is a TypeGen auto-generated file.
 * Any changes made to this file can be lost when this file is regenerated.
 */

/**
 * <para>A representation of geometry for mesh, line, or point. 
 * Includes vertex positions, face indices, normals, colors, UVs, and custom attributes within buffers, 
 * reducing the cost of passing all this data to the GPU.</para>
 * <para>Wrapper for three.js <a target="_blank" href="https://threejs.org/docs/index.html#api/en/core/BufferGeometry">BufferGeometry</a></para>
 */
export class BufferGeometry_B3D {
    /**
     * Optional name of the object. Default is an empty string. It has not to be unique.
     */
    name: string;
    /**
     * <a href="https://en.wikipedia.org/wiki/Universally_unique_identifier" target="_blank">Universal unique identifier</a> of this object instance. It's automatically assigned Guid, so it shouldn't be edited.
     */
    uuid: string;
    type: string;
}